[backport] Accessibility checks for methods with an InvokeDynamic instruction
Implements the necessary tests to check if a method with an
InvokeDynamic instruction can be inlined into a destination class.

Only InvokeDynamic instructions with LambdaMetaFactory as bootstrap
methods can be inlined. The accessibility checks cannot be implemented
generically, because it depends on what the bootstrap method is doing.
In particular, the bootstrap method receives a Lookup object as
argument which can be used to access private methods of the class
where the InvokeDynamic method is located.

A comment in the inliner explains the details.

(cherry picked from commit e9c742e1d27c41ee8646ea20dfb59efbc3d94ef3
 test removed for the 2.11.x backport as it requires Java 8
)