[indylambda] Improve test coverage
Adding tests for the selective use of boxing bridge methods
and to show that specialization is not subverted by indylambda.

Other aspects of indylambda are tested by tests like:

  - run/lambda-serialization.scala
  - run/indylambda-boxing

When those tests were written, they only tested the old backend.
However, now that we have Java 8 and the new backend avaialble
by default to partest, they provide the intended coverage.