SI-9387 Fix VerifyError introduced by indylambda
As with regular `Apply`-s, we should compute the generated type
based on the function's type, rather than the expected type.

In the test case, the expected type was void. Now, we correctly
use the generated type of `scala/Function1`, which is enough
to generate a subsequent POP instruction.

The tree shape involved was:

```
arg0 = {
  {
    $anonfun()
  };
  scala.runtime.BoxedUnit.UNIT
}
```