[backport] SI-9392 Avoid crash in GenBCode for incoherent trees
A macro in shapeless was generating a tree of the form:

```
{
	class C#2
	new C#2
}.setType(C#1)
```

This happened due to an error in the macro; it used untypecheck
to try to fix the owner-chain consistency problem, but kept a
reference to the previous version of the block-local class symbol
`C` and used this in the resulting tree.

This commit detects the particular situation we encountered, and
avoids the crash by not creating the `NestedInfo` for the
`BType` corresponding to `C#1`. The code comment discusses why I
think this is safe, and suggests a refactoring that would mean
we only ever try to construct `NestedInfo` when we are going to
need them.