SI-9396 Runner computes path only once
Change the classpath URL list in the runner settings to a lazy val.

Also clean up PathResolver's use of settings.classpath so
that the default is defined in one place, namely in settings,
where it can also be overridden.

The previous definition in both places was the same, namely,
`sys.env.getOrElse("CLASSPATH", ".")`, but the history of the
code path is fraught.