SI-9401 Avoid SOE with array + missing classtag
The implicit classtags required by the Array constructor are not
expressed in the type signature of its constructor, and instead
are summoned by a special case in the typechecker.

This special case entails replacing the `new Array` tree with
`implicitly[T].newArray(size)`, handled in `ArrayInstantiation`.
This tree is recursively typechecked.

However, if the implicit materialization/search fails, an error
is issued to the current reporter and the original tree is marked
with an error type. As above, this is recursively typechecked.

In the normal course of affairs, the recursive typecheck of the
erroneous tree would be a noop (the tree already has a type!).

However, if we are both in silent mode (in which errors are buffered)
and in retyping mode (in which the typer clears the type and symbols
of trees), we were getting into an cycle.

In the enclosed test, retyping mode was trying to recover
from:

    Resetting.this.gencastarray_=(new Array[T](0).<ERROR>)

By inserting a suitable a view:

    implicitly[Resetting => { def gencastarray_=(AT)}](
      Resetting.this
    ).gencastarray_=(new Array[T](0))

Where AT is the type found by retypechecking the argument.
It is during the argument retypechecking that we fell into cycle.

Crazily enough, in 2.11.0, the ensuing `StackOverflowError` was
being caught and treated as a failure. We would then back out of
the retyping mode, and issue the error from the the very first attempt
at the implicit search.

This fragile state of affairs was disrupted by a refactoring to
the error reporting system in 725c5c9, after which the SOE crashed
the compiler.

This commit avoids recursively typechecking error typed trees.