SI-8502 create PackageClass instead of Class symbol stubs for pkgs
https://github.com/scala/scala/pull/4111 creates a stub type symbol
for missing packages, deferring (or avoiding) a crash if a package
is missing.

The symbol created was a ClassSymbol, which could lead to an assertion
failure in flattten:

    case TypeRef(pre, sym, args) if isFlattenablePrefix(pre) =>
      assert(args.isEmpty && sym.enclosingTopLevelClass != NoSymbol, sym.ownerChain)

`pre` is the stub ClassSymbol, so `isFlattenablePrefix` is true (but
it should be false). The assertion then fails because the enclosing
class of a top-level class defined in a missing package is NoSymbol.

This failed only with GenBCode, which traverses more of the symbol
graph while building ClassBTypes: it looks collects the nested classes
of `Outer` into a `NestedInfo`.