SI-9408 Record known subclasses of local classes
Using the same facility that we use to record subclasses of
sealed classes, record the subclasses of term-owned ("local")
classes.

I have changed existing callers of `children` to use `sealedChildren`
so we don't start using this new information in pattern matching
and type pattern checkability analysis.

The following commit will build on this to infer finality of local
classes in the context of outer pointer elision in the constructors
phase.