Ignore OpenJDK warnings in partest filters
When running a 2.11.x validation on JDK 8, there are OpenJDK
warnings that fail tests. Backports a part of 8d2d3c70 to 2.11.x.

The missing filter causes merge commits to fail in PRs that merge
2.11 to 2.12, e.g., https://github.com/scala/scala/pull/4655.

In the future we probably want to trigger an integration build for
merge commits, but this doesn't happen yet AFAICT.