Cleanup in Refchecks
For historical reasons, when eliminating ModuleDef trees, RefChecks
would check if moduleVar field already exists, and only create it if
not. In reality, the lookup would always fail.

When initially committed, the moduleVar could be created either by
the RefChecks transformer or info transformer, see 256aca6.

This was later changed (3f1f0a4), after which RefChecks only creates
a moduleVar when eliminating a ModuleDef.