SI-9365 Don't null out dependencies of transient lazy vals
As per Iulian's analysis:

> When lazy vals are transient, the optimization in SI-720 is invalid,
> leading to NPEs.

These NPEs appear when recomputing the lazy val when deserializaing
the object.

This commit disables the field nulling if the lazy val is marked
transient.

The post-mixin tree in the enclosed test changes as follows:

```
--- sandbox/old.log	2015-07-27 15:48:03.000000000 +1000
+++ sandbox/new.log	2015-07-27 15:47:56.000000000 +1000
@@ -1,61 +1,54 @@
 [[syntax trees at end of                     mixin]] // t9365.scala
 package <empty> {
   class Test extends Object with Serializable {
     @transient @volatile private[this] var bitmap$trans$0: Boolean = false;
     private def foo$lzycompute(): Object = {
       {
         Test.this.synchronized({
           if (Test.this.bitmap$trans$0.unary_!())
             {
               Test.this.foo = Test.this.x.apply();
               Test.this.bitmap$trans$0 = true;
               ()
             };
           scala.runtime.BoxedUnit.UNIT
         });
-        Test.this.x = null
+        ()
       };
       Test.this.foo
     };
```

In addition to the test from the ticket, I've added a reflection
based test that directly tests the nulling. This complements the
test added in 449f2a7473, the fix for SI-720, which passes by
virtue of not exhausting the heap.