SI-3623 Improved error message for "filename too long" build errors
When building on ecryptfs filenames can be limited to ~142 characters.
This limit doesn't take long to hit and can leave the the user with a
hard to diagnosis error message.  Some legacy file systems will have
similarly small limits.  This just adds a hint that the error might
be related to the underlying fs.