SI-6756 Allow reification of A#B where A has a type tag
Before, we incurred a "macro expansion has free type B" error
at runtime, which was the result of an internal call to
`reifySym(symbolOf[B])`.

This commit instead reifies a looks up the symbol `B` as a member
of the prefix type.

Here is the diff in full of the post typer trees in the test case:

https://gist.github.com/retronym/a01032db393274d5b1dc