Unfinalize the class DefaultPromise
It was non-final in Scala 2.11.x, and made final as part
of fa0743c32.

Removing the final modifier seems like the cleanest way to enable
conversions like `javaFuture.toScala.toJava` to return the original
`javaFuture` in scala-java8-compat.

I have made the methods defined in this class final as an
alternative lockdown.

Discussion, Motivation:

  https://github.com/scala/scala-java8-compat/pull/46
  https://github.com/scala/scala-java8-compat/pull/50