Ensure the call graph contains all inlined callsites
After inlining, inlined callsites are added to the call graph.
For efficiency, the callsites of the callee are copied and adapted
(the alternative would be to re-add the entire method to the call
graph, which involves running data flow analyses).

If the callee is in a class parsed from bytecode (vs a class being
compiled), it might not yet be in the call graph. This commit ensures
that the method is added to the call graph before looking for its
callsites.