Fix maxStack after inlining
The computed value was too large previously: inlining stores receiver
and argument values of the inlined method into locals.

A too large value doesn't cause any bugs, but data flow analyses would
allocate too large frames.

The test (InlinerTest.maxLocalsMaxStackAfterInline) didn't catch the
bug because it tested the methods of the wrong ClassNode. The
CodeGenTools.compileClasses method returns a ClassNode that is created
from the classfile byte array emitted by the compiler. This ClassNode
is a new object and unrelated to the ClassNode used in the compiler
backend. The test now takes the correct ClassNode from the hash map in
the backend's byteCodeRepository.