Add an option to let pres compiler user manage threading
The presentation compiler currently demands that all interaction
is performed by asynchronous submission of work items, which are
queued and executed on the presentation compiler thread.

This is fairly inconvenient if you are a known-single-threaded client
that is trying to use the compiler from your own thread.

This commit adds an option to disable "assertCorrectThread" to better
support this use case.