SI-9429 fix filtering in scaladoc after focusing on a package
The scaladoc left pane used to be refreshed too often until ce5cfd2
when it was changed to only refresh when the filter changed. This forgot
one important case when it also must be refreshed: when entering the
"focus on a single package" mode (and it has been broken ever since).

This forces a refresh when focusing on a package, and makes sure all
entities are displayed when filtering.