remove old partest.bat batch file
this hadn't been touched since 2013, and didn't work when I tried it
-- not surprisingly since the classpath stuff for partest changed a
lot when Scala was modularized.

the shell script works on Cygwin, so the bat file isn't needed.

further piece of evidence this file is abandoned and unused: at
https://issues.scala-lang.org/browse/SI-3290, in 2010, Lukas Rytz
wrote "i never heard somebody saying it was abandoned, but i never
heard of somebody using it either."