use latest partest (1.0.9)
the new version should be no different, from this repo's perspective,
since the changes made between 1.0.7 and 1.0.9 had only to do with
build and packaging.  nonetheless, we should be using the latest
latest to help guard against regressions.

(my other motive is that I'm contemplating fixing a partest issue
that would result in a 1.0.10 release, so I'd like to have the
upgrade to 1.0.9 in place first, so if anything goes wrong there
is less searching to do for the cause)