SI-4775 Prefer non-vararg when not tupling
When the args align, the explicit test for tupling is ducked,
and simple conformance is required. But for checking
applicability, this is implicitly a tupling, so disallow it
because a vararg only conforms to a vararg.

```
// people want to do this, too
public String k(Object x) { return "foo"; }     // pick me, pick me!
public <T> String k(T... xs) { return "bar"; }
```