Support completion in erroneous string interpolation.
In the code:

```
s"${fooo<CURSOR"
```

The parser treats `fooo` as a interpolator ID for the quote that
we actually intend to end the interpolated string.

Inserting a space (in addition to `__CURSOR__` that we already
patch in to avoid parsing a partial identifier as a keyword),
solves this problem.