Improve presentation compilation of annotations
A trio of problems were hampering autocompletion of annotations.

First, given that that annotation is written before the annotated
member, it is very common to end parse incomplete code that has a
floating annotation without an anotatee.

The parser was discarding the annotations (ie, the modifiers) and
emitting an `EmptyTree`.

Second, the presetation compiler was only looking for annotations
in the Modifiers of a member def, but after typechecking annotations
are moved into the symbol.

Third, if an annotation failed to typecheck, it was being discarded
in place of `ErroneousAnnotation`.

This commit:
  - modifies the parser to uses a dummy class- or type-def tree,
    instead of EmptyTree, which can carry the annotations.
  - updates the locator to look in the symbol annotations of the
    modifiers contains no annotations.
  - uses a separate instance of `ErroneousAnnotation` for each
    erroneous annotation, and stores the original tree in its
    `original` tree.