add missing newline at eof in Scala source
this was causing a mysterious compilation failure on Windows.  (it may
not have been a sufficient cause in itself -- which is why I say
"mysterious" -- but in any case, adding the newline made the failure
go away. and besides, the newline should be there. so here it is.)

(it's tempting to make a big commit that fixes this in every
source file. resisting for now)