Don't create inline requests for callsites that cannot be inlined
When traversing the call graph and collecting inline reqeusts, rule
out callsites that we already know cannot be inlined.

Note that we cannot perform all necessary checks already at this
stage: checks that depend on the callee body (the inlined code) are
deferred until the callsite is actually inlined. The reason is that
the code may change. Example:

    @inline final def f = try 1 catch { case _: Throwable => 2 }
    @inline final def g = f
    def t = println(g)

When collecting inline requests, the body of g invokes the public
method f, so g could be inlined into t. However, once f is inlined
into g, the body of g contains a try-catch block. Now we cannot inline
g into t anymore, because the call stack at the g callsite is
non-empty (the stack is cleared when entering a handler).